fixed doubles and ints

+++ src_c/draw.c
@@ -2981,12 +2981,13 @@ draw_ellipse_xiaolinwu(SDL_Surface *surf, int x0, int y0, int width,
     int y = b;
     int layer_b = (b - thickness);
     for (int layer_a = a - thickness; layer_a <= a; layer_a++) {
-        float pow_layer_a = pow(layer_a, 2);
-        float pow_layer_b = pow(layer_b, 2);
+        double pow_layer_a = pow(layer_a, 2);
+        double pow_layer_b = pow(layer_b, 2);
         double prev_opacity = 0.0;
         x = 0;
         y = layer_b;
-        int ffd = round(pow_layer_a / sqrt(pow_layer_a + pow_layer_b)) + 1;
+        int ffd =
+            (int)round(pow_layer_a / sqrt(pow_layer_a + pow_layer_b)) + 1;
         if (layer_a == a - thickness) {
             while (x < ffd) {
                 double height = layer_b * sqrt(1 - pow(x, 2) / pow_layer_a);
@@ -3003,7 +3004,8 @@ draw_ellipse_xiaolinwu(SDL_Surface *surf, int x0, int y0, int width,
             }
             x = layer_a + 1;
             y = 0;
-            ffd = round(pow_layer_b / sqrt(pow_layer_a + pow_layer_b)) + 1;
+            ffd =
+                (int)round(pow_layer_b / sqrt(pow_layer_a + pow_layer_b)) + 1;
             while (y < ffd) {
                 double width = layer_a * sqrt(1 - pow(y, 2) / pow_layer_b);
                 double opacity = 255.0 * (ceil(width) - width);
@@ -3034,7 +3036,8 @@ draw_ellipse_xiaolinwu(SDL_Surface *surf, int x0, int y0, int width,
             }
             x = layer_a + 1;
             y = 0;
-            ffd = round(pow_layer_b / sqrt(pow_layer_a + pow_layer_b)) + 1;
+            ffd =
+                (int)round(pow_layer_b / sqrt(pow_layer_a + pow_layer_b)) + 1;
             while (y < ffd) {
                 double width = layer_a * sqrt(1 - pow(y, 2) / pow_layer_b);
                 double opacity = 255.0 * (ceil(width) - width);
@@ -3065,7 +3068,8 @@ draw_ellipse_xiaolinwu(SDL_Surface *surf, int x0, int y0, int width,
             }
             x = layer_a + 1;
             y = 0;
-            ffd = round(pow_layer_b / sqrt(pow_layer_a + pow_layer_b)) + 1;
+            ffd =
+                (int)round(pow_layer_b / sqrt(pow_layer_a + pow_layer_b)) + 1;
             while (y < ffd) {
                 double width = layer_a * sqrt(1 - pow(y, 2) / pow_layer_b);
                 double opacity = 255.0 * (ceil(width) - width);
@@ -3100,7 +3104,7 @@ draw_ellipse_xiaolinwu_thin(SDL_Surface *surf, int x0, int y0, int width,
     int y = b;
     // 45 degree coordinate, at that point switch from horizontal to vertical
     // drawing
-    int ffd = round(pow_a / sqrt(pow_a + pow_b)) + 1;
+    int ffd = (int)round(pow_a / sqrt(pow_a + pow_b)) + 1;
     while (x < ffd) {
         double height = b * sqrt(1 - pow(x, 2) / pow_a);
         double opacity = 255.0 * (ceil(height) - height);
@@ -3117,7 +3121,7 @@ draw_ellipse_xiaolinwu_thin(SDL_Surface *surf, int x0, int y0, int width,
     // vertical drawing
     x = a + 1;
     y = 0;
-    ffd = round(pow_b / sqrt(pow_a + pow_b)) + 1;
+    ffd = (int)round(pow_b / sqrt(pow_a + pow_b)) + 1;
     while (y < ffd) {
         double width = a * sqrt(1 - pow(y, 2) / pow_b);
         double opacity = 255.0 * (ceil(width) - width);
